requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
accelerate=0.31.0=pyhd8ed1ab_0
aiohttp=3.9.5=py312h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=14.0.2=h374c478_1
attrs=23.1.0=py312h06a4308_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.10.55=h721c034_0
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.7=py312ha883a20_0
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cmake=3.29.6=pypi_0
colorama=0.4.6=py312h06a4308_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
datasets=2.14.4=pyhd8ed1ab_0
diffusers=0.29.2=pyhd8ed1ab_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
einops=0.8.0=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py312h5eee18b_0
fsspec=2024.6.1=pyhff2d567_0
galois=0.3.10=pypi_0
gflags=2.2.2=h6a678d5_1
glog=0.5.0=h6a678d5_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
hamming-codec=0.3.5=pypi_0
huggingface_hub=0.23.4=pyhd8ed1ab_0
icu=73.1=h6a678d5_0
idna=3.7=py312h06a4308_0
imageio=2.34.2=pyh12aca89_0
importlib-metadata=8.0.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py312h06a4308_0
jpeg=9e=h5eee18b_1
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcurand=10.3.6.39=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=14.1.0=hc0a3c3a_0
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=h5eee18b_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.42.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.0.4=py312h5eee18b_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py312h06a4308_0
numba=0.59.1=pypi_0
numexpr=2.8.7=py312hf827012_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.1=h4ab18f5_1
orc=1.7.4=hb3bc3d3_1
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=py312h526ad5a_0
pathos=0.3.2=pypi_0
pillow=10.3.0=py312h5eee18b_0
pip=24.0=py312h06a4308_0
pox=0.3.4=pypi_0
ppft=1.7.6.8=pypi_0
prettytable=3.10.0=pypi_0
psutil=6.0.0=py312h9a8786e_0
pyarrow=14.0.2=py312hb107042_0
pybind11=2.13.1=py312h2492b07_0
pybind11-global=2.13.1=py312h2492b07_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0post0=py312h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py312h5eee18b_1
python_abi=3.12=4_cp312
pytorch=2.3.1=py3.12_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py312h06a4308_0
pyyaml=6.0.1=py312h5eee18b_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
reedmuller=1.1.2=pypi_0
regex=2024.5.15=py312h9a8786e_0
requests=2.32.2=py312h06a4308_0
s2n=1.3.27=hdbd6064_0
safetensors=0.4.3=py312h4b3b743_0
scikit-build=0.18.0=pypi_0
setuptools=69.5.1=py312h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=py312hfef1a59_0
torchaudio=2.3.1=py312_cu121
torchvision=0.18.1=py312_cu121
tqdm=4.66.4=pyhd8ed1ab_0
transformers=4.42.4=pyhd8ed1ab_0
typing-extensions=4.11.0=py312h06a4308_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=py312h06a4308_0
utf8proc=2.6.1=h5eee18b_1
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py312h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.3=py312h5eee18b_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.5=hc292b87_2
